Set up for multi-layer perceptron
Copies from perceptronMNIST.py, removes unnecessary and confusing instantiations of vectors,
renames variables, creates variables for the number of hidden units, batch size, and alpha
(the momentum constant). Loops through all epochs, regardless of accuracy achieved.
Uses sigmoid function to set activations. Sets the targets for each digit to 0.1 if not the target, 0.9 if
it is the target (in the y_target array).